// components/api/openapi/models/SirutaDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SirutaDTO = {
    code: number;
    name: string;
    fullName: string;
    parentCode?: number;
    level: number;
    localCode: number;
    mnemonic?: string;
    sirutaType: number;
    internalType: number;
    region: number;
    countryCode: string;
};